requirements.txt
py-dateutil
hypothesis
pandas
pytest
pytest-cov
